+++ src/lib/pages/codes/index.tsx
@@ -1,11 +1,4 @@
-import {
-  Heading,
-  Tabs,
-  TabList,
-  TabPanels,
-  TabPanel,
-  Flex,
-} from "@chakra-ui/react";
+import { Heading, Box, Flex, Text } from "@chakra-ui/react";
 import { observer } from "mobx-react-lite";
 import { useRouter } from "next/router";
 import type { ChangeEvent } from "react";
@@ -13,23 +6,22 @@ import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 
 import { useInternalNavigate } from "lib/app-provider";
-import { CustomTab } from "lib/components/CustomTab";
 import { FilterByPermission } from "lib/components/forms";
 import InputWithIcon from "lib/components/InputWithIcon";
 import PageContainer from "lib/components/PageContainer";
+import { EmptyState } from "lib/components/state";
+import { CodesTable } from "lib/components/table";
 import type { PermissionFilterValue } from "lib/hooks";
-import { useMyCodesData } from "lib/model/code";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 
-import { MySavedCodesSection } from "./components/MySavedCodesSection";
-import { MyStoredCodesSection } from "./components/MyStoredCodesSection";
+import { useRecentCodesData } from "./data";
 
-interface CodeFilterState {
+interface RecentCodesState {
   keyword: string;
   permissionValue: PermissionFilterValue;
 }
 
-const Codes = observer(() => {
+const RecentCodes = observer(() => {
   const router = useRouter();
   const navigate = useInternalNavigate();
   const onRowSelect = (codeId: number) =>
@@ -38,48 +30,40 @@ const Codes = observer(() => {
       query: { codeId },
     });
 
-  const { watch, setValue } = useForm<CodeFilterState>({
+  const { watch, setValue } = useForm<RecentCodesState>({
     defaultValues: {
       permissionValue: "all",
       keyword: "",
     },
   });
   const { keyword, permissionValue } = watch();
-
-  const {
-    storedCodesCount,
-    storedCodes: stored,
-    savedCodesCount,
-    savedCodes: saved,
-    allCodesCount,
-    isStoredCodesLoading,
-    isSavedCodesLoading,
-  } = useMyCodesData(keyword, permissionValue);
-
-  const isSearching = !!keyword || permissionValue !== "all";
+  const { recentCodes, isLoading } = useRecentCodesData(
+    keyword,
+    permissionValue
+  );
 
   useEffect(() => {
-    if (router.isReady) AmpTrack(AmpEvent.TO_MY_CODES);
+    if (router.isReady) AmpTrack(AmpEvent.TO_RECENT_CODES);
   }, [router.isReady]);
+
+  const isSearching = Boolean(keyword) || permissionValue !== "all";
 
   return (
     <PageContainer>
-      <Heading
-        variant="h5"
-        as="h5"
-        minH="36px"
-        display="flex"
-        alignItems="center"
-      >
-        My Codes
-      </Heading>
-      <Tabs mt={8}>
-        <TabList mb={8} borderBottom="1px" borderColor="gray.800">
-          <CustomTab count={allCodesCount}>All Codes</CustomTab>
-          <CustomTab count={storedCodesCount}>My Stored Codes</CustomTab>
-          <CustomTab count={savedCodesCount}>My Saved Codes </CustomTab>
-        </TabList>
-        <Flex gap={3} pb={4}>
+      <Box pb={12}>
+        <Heading
+          variant="h5"
+          as="h5"
+          minH="36px"
+          display="flex"
+          alignItems="center"
+        >
+          Recent Codes
+        </Heading>
+        <Text variant="body2" color="text.dark" fontWeight="500">
+          These codes are the most recently stored on this network
+        </Text>
+        <Flex gap={3} mt={8}>
           <InputWithIcon
             placeholder="Search with code ID or code name"
             value={keyword}
@@ -96,43 +80,25 @@ const Codes = observer(() => {
             }}
           />
         </Flex>
-        <TabPanels>
-          <TabPanel p={0}>
-            <MyStoredCodesSection
-              codes={stored}
-              isLoading={isStoredCodesLoading}
-              onRowSelect={onRowSelect}
-              disconnectedMessage="to see your previously uploaded and stored codes."
-              isSearching={isSearching}
-            />
-            <MySavedCodesSection
-              codes={saved}
-              isLoading={isSavedCodesLoading}
-              onRowSelect={onRowSelect}
-              isSearching={isSearching}
-            />
-          </TabPanel>
-          <TabPanel p={0}>
-            <MyStoredCodesSection
-              codes={stored}
-              isLoading={isStoredCodesLoading}
-              onRowSelect={onRowSelect}
-              disconnectedMessage="to see your previously uploaded and stored codes."
-              isSearching={isSearching}
-            />
-          </TabPanel>
-          <TabPanel p={0}>
-            <MySavedCodesSection
-              codes={saved}
-              isLoading={isSavedCodesLoading}
-              onRowSelect={onRowSelect}
-              isSearching={isSearching}
-            />
-          </TabPanel>
-        </TabPanels>
-      </Tabs>
+      </Box>
+      <CodesTable
+        codes={recentCodes}
+        isLoading={isLoading}
+        emptyState={
+          <EmptyState
+            imageVariant={isSearching ? "not-found" : "empty"}
+            message={
+              isSearching
+                ? "No matched codes found"
+                : "Most recent 100 code IDs will display here."
+            }
+            withBorder
+          />
+        }
+        onRowSelect={onRowSelect}
+      />
     </PageContainer>
   );
 });
 
-export default Codes;
+export default RecentCodes;